refactor: rename updatecart action
